单开一个工具文件，并尝试添加泛型参数
